+++ packages/media/media-viewer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/media-viewer
 
+## 13.3.1
+- [patch] use proper collectionName property in MVNG [7815256](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/7815256)
+
+## 13.3.0
+- [minor] Add collection support to Media Viewer NGwq [6baa5d0](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/6baa5d0)
+
 ## 13.2.3
 
 ## 13.2.2